rename application name
